Resultado-busqueda: Add main structure of "resultados busqueda"
