+++ src/components/Header/StyledComponents/Search/index.tsx
@@ -1,0 +1,17 @@
+import { alpha, styled } from "@mui/material/styles";
+
+const Search = styled("div")(({ theme }) => ({
+  position: "relative",
+  borderRadius: theme.shape.borderRadius,
+  backgroundColor: alpha(theme.palette.common.white, 0.15),
+  "&:hover": {
+    backgroundColor: alpha(theme.palette.common.white, 0.25),
+  },
+  width: "70vw",
+  display: "flex",
+  flexDirection: "row",
+  justifyContent: "end",
+  alignItems: "center",
+}));
+
+export default Search;
