+++ src/components/Header/StyledComponents/SearchIconWrapper/index.tsx
@@ -1,0 +1,14 @@
+import { styled } from "@mui/material/styles";
+
+const SearchIconWrapper = styled("div")(({ theme }) => ({
+  padding: theme.spacing(0, 1),
+  height: "100%",
+  position: "absolute",
+  pointerEvents: "none",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+  backgroundColor: theme.palette.secondary.main,
+}));
+
+export default SearchIconWrapper;
