+++ src/components/Header/index.tsx
@@ -1,50 +1,12 @@
 import * as React from "react";
 
-import { alpha, styled } from "@mui/material/styles";
-
 import AppBar from "@mui/material/AppBar";
 import Box from "@mui/material/Box";
-import InputBase from "@mui/material/InputBase";
+import Search from "./StyledComponents/Search";
 import SearchIcon from "@mui/icons-material/Search";
+import SearchIconWrapper from "./StyledComponents/SearchIconWrapper";
+import StyledInputBase from "./StyledComponents/StyledInputBase";
 import Toolbar from "@mui/material/Toolbar";
-
-const Search = styled("div")(({ theme }) => ({
-  position: "relative",
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  "&:hover": {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  width: "70vw",
-  display: "flex",
-  flexDirection: "row",
-  justifyContent: "end",
-  alignItems: "center",
-}));
-
-const SearchIconWrapper = styled("div")(({ theme }) => ({
-  padding: theme.spacing(0, 1),
-  height: "100%",
-  position: "absolute",
-  pointerEvents: "none",
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-  backgroundColor: theme.palette.secondary.main,
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: theme.palette.common.black,
-  width: "100%",
-  "& .MuiInputBase-input": {
-    padding: theme.spacing(1, 1, 1, 0),
-    paddingLeft: `1rem`,
-    transition: theme.transitions.create("width"),
-    width: "100%",
-    backgroundColor: theme.palette.common.white,
-    marginLeft: theme.spacing(2),
-  },
-}));
 
 function Header() {
   return (
