+++ src/components/Layout/styles.ts
@@ -1,0 +1,12 @@
+import makeStyles from "@mui/styles/makeStyles";
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    height: "100vh",
+    overflow: "hidden",
+    width: "100vw",
+    backgroundColor: "#e7e7e7",
+  },
+}));
+
+export default useStyles;
