--- src/containers/Busqueda/index.tsx
@@ -1,15 +1,0 @@
-import Header from "../../components/Header";
-import React from "react";
-import useStyles from "./styles";
-
-const Busqueda = () => {
-  const classes = useStyles();
-  return (
-    <main className={classes.root}>
-      <Header />
-      <section></section>
-    </main>
-  );
-};
-
-export default Busqueda;
