+++ src/containers/CajaBusqueda/index.tsx
@@ -1,0 +1,8 @@
+import Layout from "../../components/Layout";
+import React from "react";
+
+const CajaBusqueda = () => {
+  return <Layout id="caja-busqueda"></Layout>;
+};
+
+export default CajaBusqueda;
