+++ src/containers/ResultadosBusqueda/index.tsx
@@ -1,0 +1,10 @@
+import Layout from "../../components/Layout";
+import React from "react";
+import useStyles from "./styles";
+
+const ResultadosBusqueda = () => {
+  const classes = useStyles();
+  return <Layout id="resultados-busqueda">Resultados de búsqueda</Layout>;
+};
+
+export default ResultadosBusqueda;
