+++ src/containers/ResultadosBusqueda/styles.ts
@@ -1,11 +1,7 @@
 import makeStyles from "@mui/styles/makeStyles";
 
 const useStyles = makeStyles(() => ({
-  root: {
-    backgroundColor: "#e7e7e7",
-    height: "100vh",
-    width: "100%",
-  },
+  root: {},
 }));
 
 export default useStyles;
